Second-pass security audit: fix admin-only abilities, ACF guard, DoS caps
Findings from second security audit:

- ClearCacheAbility: require manage_options (cache flush is admin action)
- QueryActivityLogAbility: require manage_options (logs contain sensitive data)
- MachineTranslateAbility: require edit_post for posts, manage_options for strings
  (machine translation incurs API costs and creates content)
- ManageRedirectsAbility: require edit_others_posts for both list and create
  (redirect rules expose internal URL structures)
- AcfAware: validate field key against acf_get_field() before update_field()
  (prevents arbitrary meta writes through ACF path)
- TranslationAuditAbility: cap posts_per_page at 2000 (was -1, DoS risk)
- GravityFormsAbility: cast form_id to (int) in listEntries
- Plugin::clearSchemaCache: use $wpdb->prepare() with esc_like()

+++ src/Integrations/GravityForms/GravityFormsAbility.php
@@ -209,7 +209,7 @@ final class GravityFormsAbility
     public function listEntries(mixed $input = []): array|WP_Error
     {
         $input = (array) ($input ?? []);
-        $formId = $input['form_id'] ?? 0;
+        $formId = (int) ($input['form_id'] ?? 0);
         unset($input['form_id']);
 
         $response = self::restGet("/gf/v2/forms/{$formId}/entries", $input);
